refactor: split i18n namespaces
