+++ locales/app/zh_CN.json
