+++ src/renderer/src/@types/constants.ts
@@ -1,5 +1,5 @@
 export const currentSupportedLanguages = (() => {
-  const langsFiles = import.meta.glob("../../../../locales/*.json")
+  const langsFiles = import.meta.glob("../../../../locales/app/*.json")
 
   const langs = [] as string[]
   for (const key in langsFiles) {
