+++ src/renderer/src/@types/i18next.d.ts
@@ -1,9 +1,9 @@
-import type resources from "./resources"
+import type resources from "./default-resource"
 
 declare module "i18next" {
   interface CustomTypeOptions {
     resources: (typeof resources)["en"]
-    defaultNS: "translation"
+    defaultNS: "app"
     // if you see an error like: "Argument of type 'DefaultTFuncReturn' is not assignable to parameter of type xyz"
     // set returnNull to false (and also in the i18next init options)
     // returnNull: false;
