--- src/renderer/src/@types/resources.ts
@@ -1,20 +1,0 @@
-import en from "../../../../locales/en.json"
-import common_en from "../../../../locales/namespaces/common/en.json"
-import common_zhCN from "../../../../locales/namespaces/common/zh_CN.json"
-import lang_en from "../../../../locales/namespaces/lang/en.json"
-import lang_zhCN from "../../../../locales/namespaces/lang/zh_CN.json"
-import zhCN from "../../../../locales/zh_CN.json"
-
-const resources = {
-  en: {
-    translation: en,
-    lang: lang_en,
-    common: common_en,
-  },
-  zh_CN: {
-    translation: zhCN,
-    lang: lang_zhCN,
-    common: common_zhCN,
-  },
-}
-export default resources
